+++ lib/core/shapes/point.ts
@@ -1,0 +1,9 @@
+/**
+ * A point is a location in Euclidean space, represented by an ordered pair (x, y) of numbers.
+ * A point is an abstract concept that can't be drawn.
+ * A point is a 0-dimensional object.
+ * A point has no length, width, hight or depth.
+ */
+export class Point {
+  constructor(public x: number, public y: number) {}
+}