+++ lib/main.ts
@@ -1,0 +1,7 @@
+// collisions
+export * from './core/collisions-2D';
+
+// shapes
+export { Circle } from './core/shapes/circle';
+export { Point } from './core/shapes/point';
+export { Rectangle } from './core/shapes/rectangle';