+++ CHANGELOG.md
@@ -4,11 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2020-XX-XX
+---
+
+## [X.X.X] - XXXX-XX-XX
 ### Added
+### Changed
+### Fixed
+
+---
+
+## [1.0.6] - 2020-XX-XX
+### Added
+- Add Point-Point collision
+- Add Point-Point collision with buffer zone
+- Add Point-Circle collision
+- Add Circle-Circle collision
+- Add Point-Rectangle collision
+- Add following shapes: Circle, Point, Rectangle
 - Add src folder with main.ts
 - Add .gitignore, .npmignore, .prettierrc, CHANGELOG.md, CODE_OF_CONDUCT.md
 - Add LICENCE, package.json, README.md and tsconfig.json
-
-### Changed
-### Fixed